requirements.txt
selenium
pandas
webdriver_manager
tqdmpi